Remove the new string.Split(string) overload
The new `Split(string)` overload breaks source compat with uses of
`Split(null)`, which is documented to split based on white space,
because it makes the call ambiguous between `Split(char[])` and
`Split(string)`. To maintain source compatibilty, we'll remove the
new `Split(string)` overload.

That leaves us with the following new overloads:

- Split(char);
- Split(char, StringSplitOptions);
- Split(char, int, StringSplitOptions);
- Split(string, StringSplitOptions);
- Split(string, int, StringSplitOptions);