Support "offline" initializtion of Tools
This change allows init-tools to function in an "offline" mode where
tools are picked up from standalone folders. Specifically, it introduces
support for two new environment variables:

- DOTNET_TOOLSET_DIR
- BUILD_TOOLS_TOOLSET_DIR

If either is set, instead of downloading toolsets, we copy an already
existing one from the folder.  The TOOLSET_DIR is a folder with sub
directories for every version of the tool in question.

For buildtools, we expect a published toolset (sans the "dotnetcli"
folder) not just a set of nuget packages (i.e. the layout of Tools/
after running ./init-tools.sh in "online" mode).