Add LatestCommit and VersionSeedDate to build.sh
By default, the build uses git to check the SHA1 of HEAD and the date of
that commit when building. In environments where git is not preset (or
the source tree does not have git metadata, e.g. when a tarball is
downloaded from GitHub) the default logic fails.

Expose explicit arguments to build.sh so we can pass these values as
part of a build and set them in cases where we don't want to invoke
git to compute this information.