Remove dependency of System.Globalization.Native.so on specific ICU version
This change adds a shim layer between the ICU libraries and the code in
System.Globalization.Native.so that automatically picks a version of the ICU
libraries that is installed on the current machine. This enables usage of
the same System.Globalization.Native.so on various Linux distributions and
their versions. The minimal ICU version supported is U_ICU_VERSION_MAJOR_NUM,
the maximum supported is the minimum plus 20.