Abstract `gtRegNum` into RegAt{Def,Use}.
With the advent of operands that are register-optional at the point of
use, the validity of `gtRegNum` has become dependent on whether or not
an operand is in fact in a register at the point of use: if the operand
is in a register, then `gtRegNum` is the register that it occupies at
the point of use; if it is not in a register, then `gtRegNum` is the
register that the operand occupied at the point of def. Failing to
observe this distinction can result in silent bad code generation, as
was observed in VSO 278376: in this case, the spilled result of a
double-returning call that was used by a double->double cast was never
moved into the cast's destination register because the code generator
was using `gtRegNum` to determine whether or not a move was necessary.

To fix this issue, this change introduces two new APIs on GenTree:
`GenTree::RegAtDef` and `GenTree::RegAtUse`. The former returns the
register defined by a node (if any), and the latter returns the register
in which the node's value is available at the point at which it is used.
`gtRegNum` has also been changed to assert that the node is not contained,
in which case `gtRegNum` may return an invalid result for reg-optional
operands.

The new APIs should be used as follows:
- When generating code for a node, `RegAtDef()` should be used to
  determine the register that will contain the node's value
- When generating code that consumes a particular node, if the node may be
  contained (either because it was marked as such during lowering or
  because it is register-optional), `RegAtUse()` should be used to
  determine the register in which the node's value is available. If the node
  may not be a memory operand, `gtRegNum` should be used instead.