Fix assert with struct copy to/from same lclVar
This was a case where we reused a struct temp, and then it wound up
eventually getting assigned the value it originally contained,
so we had V09 = V09.
This hit the assert in liveness where it wasn't expecting to find a
struct assignment with the same lclVar on the lhs and rhs.
This assert should have been eliminated with the IR change to struct assignments,
but when this situation arises we may run into issues if we call a helper
that doesn't expect the lhs and rhs to be the same.
So, change morph to eliminate such assignments.

Fix #7954

Fix release build