Allow creating SortKey with Upper case first sorting attribute
Currently we don't have a way to sort to creat sort keys with the option upper case first. System.Xml.Xsl support transforming Xml docs using Xslt template with the option
      <xsl:sort select = \"firstname\" case-order=\"upper-first\"/>
which is sorting elements with uppercase first.

On Windows, System.Xml.Xsl support this option by creating the sort key (without upper-case first as it is not supported) and then manually manipulate the sort key bytes to make
the result look like it is sorted as upper-case first. this will not work nicely on Linux because the sortkeys there are different than the sortkeys generated on Windows and it
is not right to manually manipulate the sort keys after creating it.

The solution here is to allow passing some flag to sort key creation when running on Linux and then use this flag to instruct ICU to generate the sort key with upper-case first option