Admit block nodes in `GenTree::AsOp`.
All block nodes are subtypes of `GenTreeOp`, but not all block nodes are
simple nodes: in particular, DYN_BLK and STORE_DYN_BLK are special
nodes. Because these nodes are not considered "simple", `GenTree::AsOp`
currently asserts when called on such a node. This change relaxes the
assertion to admit block nodes as well as simple nodes.

Fixes VSO 289704.