Avoid storing successfully completed tasks in Task.WhenAll
When WhenAll is used with an enumerable, such as in:
```C#
await Task.WhenAll(inputs.Select(async input =>
{
    ...
}));
```
it's not uncommon for some of the tasks to already be completed by the time WhenAll sees them.  Currently all of the tasks are stored and counted in the countdown list that tracks which tasks remain.  We can instead easily filter those tasks out before they're added to the list, avoiding those overheads (e.g. interlocked operations), the extra array size, unnecessarily keeping those objects alive, etc.