Update PATENTS.TXT
Rationale: The "as part of" conditions don't make sense, rely on an impossible-to-evaluate ".NET Runtime" definition, and — I'm told by people who should know — were not intended to limit the scope of the license. More info at #375

Based on that information, I suggest the removal of any mention of the .NET Runtime, and the replacement of these unintentional conditions with an an unambiguous "Covered Code, a portion thereof, or a derivative work thereof.", which should cover all use cases that the original version intended (but failed) to include with the "as  part of" phrasing. 

Fixes #375

There is also a nitpick in the definition of ".NET Patents", as it unconditionally excludes from this promise any patent claims infringed by "Enabling Code", which would be alright — except the definition thereof includes "applications that run on the .NET Framework". It's unlikely that the intent here could be misconstrued, but there is shorter, simpler and clearly way to say this: 

".NET Patents" are those patent claims, both currently owned by Microsoft and 
acquired in the future, that are necessarily infringed by Covered Code. .NET 
Patents do not include any patent claims that are infringed only as a 
consequence of modification of Covered Code, or that are infringed 
only by the combination of Covered Code with Enabling Technology or third party code.