Fix NULL string printing in printf
The functions implementing various printf flavors in PAL were not handling
correctly the case when a string pointer passed to them for %S and %s formats
was NULL. Instead of printing `(null)`, they were crashing.

This change fixes the problem and also adds PAL tests to verify it is working
properly.