Treat array indices as unsigned
Array indices are signed according to the ECMA spec but they're subject to bound checks that reject negative values so we can treat them as unsigned when widening to pointer size.

The advantage is that we end up using a `mov` instruction instead of a `movsx` instruction. The former sometimes has smaller encoding and it is supposed to have 0 latency on recent Intel CPUs.