Remove lazy initialization of Task.CompletedTask
The lazy-initialization code of Task.CompletedTask is causing it to be an "unprofitable inline".  For workloads that make heavy use of Task-returning methods that complete synchronously, this is showing up as a measurable impact, e.g. in a simple benchmark reading and writing to the channels in the channels lib on corefxlabs, it's resulting in 2-5% overhead.  However, as it's used by a bunch of core .NET types, e.g. a bunch of streams like FileStream and MemoryStream, it's rare that it won't be needed, and it's often needed very early in a program's execution.  This commit just removes the lazy-initialization, removing a branch and allowing the getter to be inlined.